requirements.txt
absl-py==0.7.1
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
category-encoders==1.3.0
cycler==0.10.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
more-itertools==7.0.0
numpy==1.16.2
pandas==0.24.2
pathlib2==2.3.3
patsy==0.5.1
pbr==5.1.3
pkg-resources==0.0.0
pluggy==0.9.0
protobuf==3.7.1
py==1.8.0
pydot==1.4.1
pyparsing==2.4.0
pytest==4.4.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
statsmodels==0.9.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.15.2
